+++ Changelog.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+## New Features
+
+- [Add `HandlerFuncType` trait](https://github.com/rust-osdev/x86_64/pull/439)
+
 # 0.14.11 – 2022-09-15
 
 ## New Features